+++ packages/graphqlgen/benchmarks/integration/basic/models.ts
