+++ packages/graphqlgen/benchmarks/integration/basic/schema.graphql
