+++ packages/graphqlgen/benchmarks/integration/complex/models.ts
