+++ packages/graphqlgen/benchmarks/integration/complex/prisma.graphql
