+++ packages/graphqlgen/benchmarks/integration/complex/schema.graphql
